CPU delot celosno implementiran i debagiran so templates

+++ src/abc_gpu.cuh
@@ -2,21 +2,22 @@
 #include <cuda_runtime.h>
 #include <curand_kernel.h>
 #include "problems/problems.h"
+#include "abc_main.cuh"
 
 namespace gpu{
 
+	//TODO: ova treba da go stavam vo headerot
 	//template<unsigned int DIMENSIONS>
 	void launch_abc(
-		double*      coordinates,
-		double*      values,
+		float*       coordinates,
+		float*       values,
 		int          num_of_bees, 
 		int          max_generations, 
 		int          trials_limit, 
 		opt_func     optimization_function,
-		double       lower_bounds[],
-		double       upper_bounds[],
+		float        lower_bounds[],
+		float        upper_bounds[],
 		int          steps
 	);
 
 }
-
