+++ src/problems/gpu/problems_many_local_minima.cu
@@ -1,11 +1,11 @@
 #include "problems_many_local_minima.cuh"
 #include <math.h>
 
-__device__ double problems::gpu::ackley(double* args, int n, double a, double b, double c)
+__device__ float problems::gpu::ackley(float* args, int n, float a, float b, float c)
 {
-	double result;
-	double sum_x2 = 0;
-	double sum_cx = 0;
+	float result;
+	float sum_x2 = 0;
+	float sum_cx = 0;
 	for(int i = 0; i < n; i++)
 	{
 		sum_x2 += args[i]*args[i];
@@ -17,13 +17,13 @@ __device__ double problems::gpu::ackley(double* args, int n, double a, double b,
 	return result;
 }
 
-__device__ double problems::gpu::ackley2(double* args, int n)
+__device__ float problems::gpu::ackley2(float* args, int n)
 {
 	return ackley(args, n, 20, 0.2, 2*M_PI);
 }
 
 ///\note this function is only a 2d function
-__device__ double problems::gpu::bukin6(double* args, int n)
+__device__ float problems::gpu::bukin6(float* args, int n)
 {
 	if(n < 2) return 0;
 
@@ -32,7 +32,7 @@ __device__ double problems::gpu::bukin6(double* args, int n)
 }
 
 ///\note this function in only a 2d function
-__host__ __device__ double problems::gpu::cross_in_tray(double* args, int n)
+__host__ __device__ float problems::gpu::cross_in_tray(float* args, int n)
 {
 	if(n < 2) return 0;
 
@@ -43,18 +43,18 @@ __host__ __device__ double problems::gpu::cross_in_tray(double* args, int n)
 }
 
 ///\note this function in only a 2d function
-__device__ double problems::gpu::drop_wave(double* args, int n)
+__device__ float problems::gpu::drop_wave(float* args, int n)
 {
 	if(n < 2) return 0;
 
-	double tmp = args[0]*args[0] + args[1]*args[1];
+	float tmp = args[0]*args[0] + args[1]*args[1];
 
 	return (-1 - std::cos(12*std::sqrt(tmp))) /
 		   (0.5*tmp + 2);
 }
 
 ///\note this function in only a 2d function
-__device__ double problems::gpu::eggholder(double* args, int n)
+__device__ float problems::gpu::eggholder(float* args, int n)
 {
 	if(n < 2) return 0;
 
@@ -63,19 +63,19 @@ __device__ double problems::gpu::eggholder(double* args, int n)
 }
 
 ///\note this function is only a 1d function
-__device__ double problems::gpu::gramacy_and_lee(double* args, int n)
+__device__ float problems::gpu::gramacy_and_lee(float* args, int n)
 {
 	if(n < 1) return 0;
 
-	double tmp = args[0] - 1;
+	float tmp = args[0] - 1;
 	return std::sin(10*M_PI*args[0])/2*args[0]+
 		   tmp * tmp * tmp * tmp;
 }
 
-__device__ double problems::gpu::griewank(double* args, int n)
+__device__ float problems::gpu::griewank(float* args, int n)
 {
-	double result  = 0;
-	double product = 1;
+	float result  = 0;
+	float product = 1;
 	for(int i = 0; i < n; i++)
 	{
 		result += args[i]*args[i];
@@ -87,7 +87,7 @@ __device__ double problems::gpu::griewank(double* args, int n)
 }
 
 ///\note This only a 2d function
-__device__ double problems::gpu::holder_table(double* args, int n)
+__device__ float problems::gpu::holder_table(float* args, int n)
 {
 	if(n < 2) return 0;
 
@@ -96,15 +96,15 @@ __device__ double problems::gpu::holder_table(double* args, int n)
 }
 
 ///\note A has dimensions nxm
-__device__ double problems::gpu::langerman(double* args, int n, double* c, int m, double* A)
+__device__ float problems::gpu::langerman(float* args, int n, float* c, int m, float* A)
 {
-	double result = 0;
+	float result = 0;
 	for(int i = 0; i < m; i++)
 	{
-		double sum1 = 0;
+		float sum1 = 0;
 		for(int j = 0; j < n; j++)
 		{
-			double tmp = args[j] - A[i*n+j];
+			float tmp = args[j] - A[i*n+j];
 			sum1 += tmp * tmp;
 		}
 		result += c[i] * std::pow(M_E, -1.0/M_PI*sum1) *
@@ -114,26 +114,26 @@ __device__ double problems::gpu::langerman(double* args, int n, double* c, int m
 	return result;
 }
 
-__device__ double problems::gpu::langerman2(double* args, int n)
+__device__ float problems::gpu::langerman2(float* args, int n)
 {
-	double A[] = {
+	float A[] = {
 		3, 5,
 		5, 2,
 		2, 1,
 		1, 4,
 		7, 9
 	};
-	double c[] = {1, 2, 5, 2, 3};
+	float c[] = {1, 2, 5, 2, 3};
 
 	return langerman(args, n, c, 5, A);
 }
 
-__device__ double problems::gpu::levy(double* args, int n)
+__device__ float problems::gpu::levy(float* args, int n)
 {
-	double tmp, tmp1;
-	double result = 0;
+	float tmp, tmp1;
+	float result = 0;
 
-	double w = 1.0 + (args[0]-1.0)/4.0;
+	float w = 1.0 + (args[0]-1.0)/4.0;
 	tmp = std::sin(M_PI*w);
 	result += tmp*tmp;
 
@@ -154,15 +154,15 @@ __device__ double problems::gpu::levy(double* args, int n)
 }
 
 ///\note This is only a 2d function
-__device__ double problems::gpu::levy13(double* args, int n)
+__device__ float problems::gpu::levy13(float* args, int n)
 {
 	if(n < 2) return 0;
 
-	double tmp1 = std::sin(3*M_PI*args[0]);
-	double tmp2 = args[0] - 1.0;
-	double tmp3 = std::sin(3*M_PI*args[1]);
-	double tmp4 = args[1] - 1.0;
-	double tmp5 = std::sin(2*M_PI*args[1]);
+	float tmp1 = std::sin(3*M_PI*args[0]);
+	float tmp2 = args[0] - 1.0;
+	float tmp3 = std::sin(3*M_PI*args[1]);
+	float tmp4 = args[1] - 1.0;
+	float tmp5 = std::sin(2*M_PI*args[1]);
 
 	return tmp1*tmp1 + tmp2*tmp2*(1 + tmp3*tmp3) +
 		   tmp4*tmp4*(1 + tmp5*tmp5);
@@ -171,7 +171,7 @@ __device__ double problems::gpu::levy13(double* args, int n)
 ///\brief This is a fucntion with mutiple global
 ///       maxima which is commonly used to test
 ///       optimization functions
-__device__ double problems::gpu::rastrigin(int a, double* x, int n)
+__device__ float problems::gpu::rastrigin(int a, float* x, int n)
 {
 	float result = a * n;
 	for(int i = 0; i < n; i++)
@@ -184,38 +184,38 @@ __device__ double problems::gpu::rastrigin(int a, double* x, int n)
 ///brief This function is a wrapper around rastrigin
 ///      in order to make it compatible to the problem
 ///      definition
-__device__ double problems::gpu::rastrigin2(double* args, int n)
+__device__ float problems::gpu::rastrigin2(float* args, int n)
 {
 	return rastrigin(10, args, n);
 }
 
 ///\note This is only a 2d function
-__device__ double problems::gpu::schaffer2(double* args, int n)
+__device__ float problems::gpu::schaffer2(float* args, int n)
 {
 	if(n < 2) return 0;
 
-	double x2 = args[0]*args[0];
-	double y2 = args[1]*args[1];
-	double tmp1 = std::sin(x2 - y2);
-	double tmp2 = 1.0 + 0.001*(x2+y2);
+	float x2 = args[0]*args[0];
+	float y2 = args[1]*args[1];
+	float tmp1 = std::sin(x2 - y2);
+	float tmp2 = 1.0 + 0.001*(x2+y2);
 	return 0.5 + (tmp1*tmp1 - 0.5)/tmp2/tmp2;
 }
 
 ///\note This is only a 2d function
-__device__ double problems::gpu::schaffer4(double* args, int n)
+__device__ float problems::gpu::schaffer4(float* args, int n)
 {
 	if(n < 2) return 0;
 
-	double x2 = args[0]*args[0];
-	double y2 = args[1]*args[1];
-	double tmp1 = std::cos(std::sin(std::abs(x2 - y2)));
-	double tmp2 = 1.0 + 0.001*(x2+y2);
+	float x2 = args[0]*args[0];
+	float y2 = args[1]*args[1];
+	float tmp1 = std::cos(std::sin(std::abs(x2 - y2)));
+	float tmp2 = 1.0 + 0.001*(x2+y2);
 	return 0.5 + (tmp1*tmp1 - 0.5)/tmp2/tmp2;
 }
 
-__device__ double problems::gpu::schwefel(double* args, int n)
+__device__ float problems::gpu::schwefel(float* args, int n)
 {
-	double result = 418.9829*n;
+	float result = 418.9829*n;
 	for(int i = 0; i < n; i++)
 	{
 		result -= args[i]*std::sin(std::sqrt(std::abs(args[i])));
@@ -225,11 +225,11 @@ __device__ double problems::gpu::schwefel(double* args, int n)
 }
 
 ///\note This is only a 2d function
-__device__ double problems::gpu::shubert(double* args, int n)
+__device__ float problems::gpu::shubert(float* args, int n)
 {
 	if(n < 2) return 0;
-	double sum1 = 0;
-	double sum2 = 0;
+	float sum1 = 0;
+	float sum2 = 0;
 	for(int i = 0; i < 5; i++)
 	{
 		sum1 += i*std::cos(args[0]*(i+1)+i);
