+++ src/problems/gpu/problems_steep_ridges.cuh
@@ -4,11 +4,11 @@
 
 namespace problems::gpu
 {
-	__device__ double dejong5(double* args, int n);
+	__device__ float dejong5(float* args, int n);
 
 	///\note This is only a 2d function
-	__device__ double easom(double* args, int n);
+	__device__ float easom(float* args, int n);
 
-	__device__ double michalewicz(double* args, int n, int m);
-	__device__ double michalewicz2(double* args, int n);
+	__device__ float michalewicz(float* args, int n, int m);
+	__device__ float michalewicz2(float* args, int n);
 }
